+++ src/components/App.js
@@ -2,9 +2,8 @@ import { useCachedDataQuery } from '@dhis2/analytics'
 import { useDataEngine, useDataMutation } from '@dhis2/app-runtime'
 import { CssVariables } from '@dhis2/ui'
 import cx from 'classnames'
-import PropTypes from 'prop-types'
 import React, { useState, useEffect, useRef } from 'react'
-import { connect, useDispatch } from 'react-redux'
+import { useSelector, useDispatch } from 'react-redux'
 import { tSetCurrent } from '../actions/current.js'
 import {
     acClearAll,
@@ -13,13 +12,11 @@ import {
     acSetVisualizationLoading,
 } from '../actions/loader.js'
 import { acAddMetadata, tSetInitMetadata } from '../actions/metadata.js'
-import { tAddSettings } from '../actions/settings.js'
 import {
     acSetUiFromVisualization,
     acAddParentGraphMap,
     acSetShowExpandedLayoutPanel,
 } from '../actions/ui.js'
-import { acSetUser } from '../actions/user.js'
 import { acSetVisualization } from '../actions/visualization.js'
 import { EVENT_TYPE } from '../modules/dataStatistics.js'
 import {
@@ -29,6 +26,7 @@ import {
 } from '../modules/error.js'
 import history from '../modules/history.js'
 import { getDynamicTimeDimensionsMetadata } from '../modules/metadata.js'
+import { SYSTEM_SETTINGS_DIGIT_GROUP_SEPARATOR } from '../modules/systemSettings.js'
 import { getParentGraphMapFromVisualization } from '../modules/ui.js'
 import { transformVisualization } from '../modules/visualization.js'
 import { sGetCurrent } from '../reducers/current.js'
@@ -104,32 +102,21 @@ const dataStatisticsMutation = {
     type: 'create',
 }
 
-const App = ({
-    current,
-    addMetadata,
-    addParentGraphMap,
-    addSettings,
-    clearAll,
-    clearLoadError,
-    error,
-    isLoading,
-    setCurrent,
-    setInitMetadata,
-    setLoadError,
-    setVisualization,
-    setVisualizationLoading,
-    setUiFromVisualization,
-    setUser,
-    showDetailsPanel,
-}) => {
+const App = () => {
     const dataEngine = useDataEngine()
     const [data, setData] = useState()
     const [fetchError, setFetchError] = useState()
-    const { currentUser, userSettings } = useCachedDataQuery()
     const [previousLocation, setPreviousLocation] = useState()
     const [initialLoadIsComplete, setInitialLoadIsComplete] = useState(false)
     const [postDataStatistics] = useDataMutation(dataStatisticsMutation)
     const dispatch = useDispatch()
+    const current = useSelector(sGetCurrent)
+    const isLoading = useSelector(sGetIsVisualizationLoading)
+    const error = useSelector(sGetLoadError)
+    const showDetailsPanel = useSelector(sGetUiShowDetailsPanel)
+    const { systemSettings, rootOrgUnits } = useCachedDataQuery()
+    const digitGroupSeparator =
+        systemSettings[SYSTEM_SETTINGS_DIGIT_GROUP_SEPARATOR]
 
     const interpretationsUnitRef = useRef()
     const onInterpretationUpdate = () => {
@@ -139,9 +126,22 @@ const App = ({
     useEffect(() => {
         if (!error && fetchError) {
             if (fetchError.details?.httpStatusCode === 404) {
-                clearAll(visualizationNotFoundError())
+                dispatch(
+                    acClearAll({
+                        error: visualizationNotFoundError(),
+                        digitGroupSeparator,
+                        rootOrgUnits,
+                    })
+                )
             } else {
-                clearAll(fetchError.details.message || genericServerError())
+                dispatch(
+                    acClearAll({
+                        error:
+                            fetchError.details.message || genericServerError(),
+                        digitGroupSeparator,
+                        rootOrgUnits,
+                    })
+                )
             }
         }
     }, [error, fetchError])
@@ -155,7 +155,7 @@ const App = ({
 
     const loadVisualization = async (location) => {
         setFetchError(undefined)
-        setVisualizationLoading(true)
+        dispatch(acSetVisualizationLoading(true))
         const isExisting = location.pathname.length > 1
         if (isExisting) {
             // /currentAnalyticalObject
@@ -173,9 +173,14 @@ const App = ({
                 setFetchError(error)
             }
         } else {
-            clearAll()
-            //const digitGroupSeparator = sGetSettingsDigitGroupSeparator(getState())
-            setVisualizationLoading(false)
+            dispatch(
+                acClearAll({
+                    error: null,
+                    digitGroupSeparator,
+                    rootOrgUnits,
+                })
+            )
+            dispatch(acSetVisualizationLoading(false))
         }
 
         dispatch(acSetShowExpandedLayoutPanel(!isExisting))
@@ -199,28 +204,17 @@ const App = ({
             {}
         )
 
-        addMetadata(itemsMetadata)
-        setVisualizationLoading(false)
+        dispatch(acAddMetadata(itemsMetadata))
+        dispatch(acSetVisualizationLoading(false))
 
         if (!response.rows?.length) {
-            setLoadError(emptyResponseError())
+            dispatch(acSetLoadError(emptyResponseError()))
         }
     }
 
     useEffect(() => {
-        const onMount = async () => {
-            await addSettings(userSettings)
-
-            setUser({
-                ...currentUser,
-                uiLocale: userSettings.uiLocale,
-            })
-
-            setInitMetadata()
-            loadVisualization(history.location)
-        }
-
-        onMount()
+        dispatch(tSetInitMetadata(rootOrgUnits))
+        loadVisualization(history.location)
 
         const unlisten = history.listen(({ location }) => {
             const isSaving = location.state?.isSaving
@@ -245,7 +239,7 @@ const App = ({
 
     useEffect(() => {
         if (data?.eventVisualization) {
-            setInitMetadata()
+            dispatch(tSetInitMetadata(rootOrgUnits))
 
             const { program, programStage } = data.eventVisualization
             const visualization = transformVisualization(
@@ -260,12 +254,16 @@ const App = ({
                 metadata[programStage.id] = programStage
             }
 
-            addParentGraphMap(getParentGraphMapFromVisualization(visualization))
-            setVisualization(visualization)
-            setCurrent(visualization)
-            setUiFromVisualization(visualization, metadata)
+            dispatch(
+                acAddParentGraphMap(
+                    getParentGraphMapFromVisualization(visualization)
+                )
+            )
+            dispatch(acSetVisualization(visualization))
+            dispatch(tSetCurrent(visualization))
+            dispatch(acSetUiFromVisualization(visualization, metadata))
             postDataStatistics({ id: visualization.id })
-            clearLoadError()
+            dispatch(acClearLoadError())
         }
     }, [data])
 
@@ -358,45 +356,4 @@ const App = ({
     )
 }
 
-const mapStateToProps = (state) => ({
-    current: sGetCurrent(state),
-    isLoading: sGetIsVisualizationLoading(state),
-    showDetailsPanel: sGetUiShowDetailsPanel(state),
-    error: sGetLoadError(state),
-})
-
-const mapDispatchToProps = {
-    addMetadata: acAddMetadata,
-    addParentGraphMap: acAddParentGraphMap,
-    addSettings: tAddSettings,
-    clearAll: acClearAll,
-    clearLoadError: acClearLoadError,
-    setCurrent: tSetCurrent,
-    setInitMetadata: tSetInitMetadata,
-    setVisualization: acSetVisualization,
-    setUser: acSetUser,
-    setUiFromVisualization: acSetUiFromVisualization,
-    setVisualizationLoading: acSetVisualizationLoading,
-    setLoadError: acSetLoadError,
-}
-
-App.propTypes = {
-    addMetadata: PropTypes.func,
-    addParentGraphMap: PropTypes.func,
-    addSettings: PropTypes.func,
-    clearAll: PropTypes.func,
-    clearLoadError: PropTypes.func,
-    current: PropTypes.object,
-    error: PropTypes.object,
-    isLoading: PropTypes.bool,
-    setCurrent: PropTypes.func,
-    setInitMetadata: PropTypes.func,
-    setLoadError: PropTypes.func,
-    setUiFromVisualization: PropTypes.func,
-    setUser: PropTypes.func,
-    setVisualization: PropTypes.func,
-    setVisualizationLoading: PropTypes.func,
-    showDetailsPanel: PropTypes.bool,
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(App)
+export default App